Extract TestKitSpec from PetSpec. Start MarkovPluginSpec implementation.
